pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "abbasta(fmamm)", email = "abbasta@abbasta.com"},
]
dependencies = [
    "ortools==9.7.2996",
    "matplotlib==3.8.0",
    "numpy==1.26.0",
    "Pillow==10.0.1",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}
